requirements.txt
flask
requests
maxcube-api

